move prisma client to server only move widget function to client lib
