+++ src/routes/dashboard/+page.server.ts
@@ -4,8 +4,8 @@ import { ExerciseRepo } from '$lib/exercise';
 import { ExerciseEventRepo } from '$lib/exerciseEvent';
 import { JournalEntryRepo } from '$lib/journalEntry';
 import { MetricRepo } from '$lib/metric';
-import { prisma } from '$lib/prisma';
 import { ProfileRepo } from '$lib/profile';
+import { prisma } from '$lib/server/prisma';
 import { TrainingCycleRepo } from '$lib/trainingCycle';
 import { TrainingProgramRepo } from '$lib/trainingProgram';
 import { getSessionOrRedirect } from '$lib/utils';
