+++ src/routes/journalEntry/+page.server.ts
@@ -1,7 +1,7 @@
 import { APIError } from '$lib/errors';
 import { SERVER_ERROR } from '$lib/helperTypes';
 import { JournalEntryRepo } from '$lib/journalEntry';
-import { prisma } from '$lib/prisma';
+import { prisma } from '$lib/server/prisma';
 import { getSessionOrRedirect } from '$lib/utils';
 import { journalEntrySchema } from '$lib/zodSchemas';
 import { error, fail } from '@sveltejs/kit';
