+++ src/routes/journalEntry/[id]/+page.server.ts
@@ -1,5 +1,5 @@
 import { JournalEntryRepo } from '$lib/journalEntry';
-import { prisma } from '$lib/prisma';
+import { prisma } from '$lib/server/prisma';
 import { getSessionOrRedirect } from '$lib/utils';
 import { journalEntrySchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
