+++ src/routes/profile/+page.server.ts
@@ -2,8 +2,8 @@ import { deleteFile, getSignedUrlsAndMetadata, uploadFile } from '$lib/aws/s3';
 import { ExerciseEventRepo } from '$lib/exerciseEvent';
 import { JournalEntryRepo } from '$lib/journalEntry';
 import { MetricRepo } from '$lib/metric';
-import { prisma } from '$lib/prisma';
 import { ProfileRepo } from '$lib/profile';
+import { prisma } from '$lib/server/prisma';
 import { getSessionOrRedirect } from '$lib/utils';
 import { fileUploadSchema, profileSchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
