+++ src/routes/project/+page.server.ts
@@ -1,5 +1,5 @@
-import { prisma } from '$lib/prisma';
 import { ProjectRepo } from '$lib/project';
+import { prisma } from '$lib/server/prisma';
 import { getSessionOrRedirect } from '$lib/utils';
 import { projectSchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
