+++ src/routes/trainingCycle/[id]/edit/+page.server.ts
@@ -1,5 +1,5 @@
 import { ExerciseRepo } from '$lib/exercise';
-import { prisma } from '$lib/prisma';
+import { prisma } from '$lib/server/prisma';
 import { TrainingCycleRepo } from '$lib/trainingCycle';
 import { getSessionOrRedirect } from '$lib/utils';
 import type { Crumb } from 'svelte-breadcrumbs';
