+++ src/routes/trainingProgram/[id]/scheduledSlot/[scheduledSlotId]/+page.server.ts
@@ -1,4 +1,4 @@
-import { prisma } from '$lib/prisma';
+import { prisma } from '$lib/server/prisma';
 import { TrainingProgramRepo } from '$lib/trainingProgram';
 import { getSessionOrRedirect } from '$lib/utils';
 import { trainingProgramScheduledSlotSchema } from '$lib/zodSchemas';
